requirements.txt
# LlamaIndex Workflows (Plain implementation)
llama-index-workflows
llama-cloud-services

# LangChain (Agent implementation)
langchain==1.0.2
langchain-openai==1.0.1
langchain-core==1.0.1
langchain-community==0.4.0

# Shared dependencies
jsonschema
openai
python-dotenv
streamlit
openpyxl